+++ antix-building/Antix.Building/Builder.cs
@@ -6,11 +6,23 @@ namespace Antix.Building
     public sealed class Builder<T>
         : BuilderBase<Builder<T>, T>
     {
-        public Builder()
+        public Builder(Func<T> create, Action<T> validate)
+            : base(create, validate)
         {
         }
 
-        public Builder(Func<T> create) : base(create)
+        public Builder()
+            : this(null, null)
+        {
+        }
+
+        public Builder(Func<T> create)
+            : this(create, null)
+        {
+        }
+
+        public Builder(Action<T> validate)
+            : this(null, validate)
         {
         }
     }