+++ ReleaseNotes.md
@@ -7,3 +7,7 @@
 
 ## 1.0.0
 Release - no changes to b01
+
+## 1.1.0-b01
+1. Added Validate, called after all withs
+2. Create and Validate no available as overrides on BuilderBase<>